First Commit
Working on spawning of rings (creating a new ring in the first if
statement but able to call it late in time). Maybe have an array of
rings and then add and remove new rings you make? can call it checking
if any of the rings in the array is stopped, tapped, etc.)